# openapi.yaml
openapi: 3.0.3
info:
  title: GPS FAMILIA API
  version: 2.11.0
  description: Family location, encrypted messaging, friends, marketplace, wallet, feed, rooms, and WebRTC calling.
  license:
    name: ISC
    url: https://opensource.org/licenses/ISC
servers:
  - url: /
security:
  - bearerAuth: []
paths:
...
  /api/friends/request:
    post:
      summary: Send a friend request
      responses:
        "200": { description: requested or already friends }
        "403": { description: blocked or not accepting }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer